% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcd.R
\name{bcd}
\alias{bcd}
\title{bcd}
\usage{
bcd(y, X, lambda1, lambda2, beta = rep(0, ncol(X)))
}
\arguments{
\item{y}{something}

\item{X}{something else}

\item{lambda1}{something else}

\item{lambda2}{something else}

\item{beta}{something else}
}
\value{
beta (describe beta)
}
\description{
does something
}
\examples{
# Give an example here that can run independently

}
