% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddn.R
\name{ddn}
\alias{ddn}
\title{ddn}
\usage{
ddn(data1, data2, lambda1, lambda2)
}
\arguments{
\item{data1}{describe}

\item{data2}{describe}

\item{lambda1}{something else}

\item{lambda2}{something else}
}
\value{
beta.matrix (describe beta.matrix)
}
\description{
ddn
}
\examples{
# Give an example here that can run independently

}
