% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve2d.R
\name{solve2d}
\alias{solve2d}
\title{Solve 2d}
\usage{
solve2d(y, x1, x2, lambda1, lambda2)
}
\arguments{
\item{y}{something}

\item{x1}{something else}

\item{x2}{something else}

\item{lambda1}{something else}

\item{lambda2}{something else}
}
\value{
beta (describe beta)
}
\description{
Solves 2d
}
\examples{
# Give an example here that can run independently

}
